Updated proto comments to allow better specification document generation
