+++ mlmodel/docs/sections/DataStructuresAndFeatureTypes.rst
@@ -1,0 +1,6 @@
+Data Structures & Feature Types
+===============================
+
+.. include:: ../reference/DataStructures.rst
+
+.. include:: ../reference/FeatureTypes.rst
