+++ mlmodel/docs/sections/FeatureEngineering.rst
@@ -1,0 +1,18 @@
+Feature Engineering
+===================
+
+.. include:: ../reference/Imputer.rst
+
+.. include:: ../reference/Scaler.rst
+
+.. include:: ../reference/Normalizer.rst
+
+.. include:: ../reference/OneHotEncoder.rst
+
+.. include:: ../reference/CategoricalMapping.rst
+
+.. include:: ../reference/FeatureVectorizer.rst
+
+.. include:: ../reference/DictVectorizer.rst
+
+.. include:: ../reference/ArrayFeatureExtractor.rst
