+++ mlmodel/docs/sections/GLM.rst
@@ -1,0 +1,7 @@
+Generalized Linear Models
+=========================
+
+.. include:: ../reference/GLMClassifier.rst
+
+.. include:: ../reference/GLMRegressor.rst
+
