+++ mlmodel/docs/sections/Model.rst
@@ -1,0 +1,7 @@
+Models
+======
+
+.. include:: ../reference/Model.rst
+
+.. include:: ../reference/Identity.rst
+
