+++ mlmodel/docs/sections/NeuralNetwork.rst
@@ -1,0 +1,4 @@
+Neural Networks
+===============
+
+.. include:: ../reference/NeuralNetwork.rst
