+++ mlmodel/docs/sections/SVM.rst
@@ -1,0 +1,4 @@
+Support Vector Machines
+=======================
+
+.. include:: ../reference/SVM.rst
