+++ mlmodel/docs/sections/TreeEnsembles.rst
@@ -1,0 +1,4 @@
+Tree Ensembles
+==============
+
+.. include:: ../reference/TreeEnsemble.rst
